Add protocol registration functionality for Windows and Linux
