+++ go.mod
@@ -2,4 +2,8 @@ module InventoryTPrinter
 
 go 1.23.5
 
-require golang.org/x/sys v0.29.0 // indirect
+require (
+	github.com/go-resty/resty/v2 v2.16.4 // indirect
+	golang.org/x/net v0.33.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+)
